Properly get the name of the test executable by parsing meson target data
Uses 'meson introspect --targets' to find the test executable that
contains the source file in the currently open buffer.

This fixes an issue with the existing implementation that assumed the
name of the test executable from the name of the buffer.

However this is not bullet proof. It is possible that the source file is
used in multiple test executables, for instance to run with different
parameters and flags.